Fix error on opening the discussion page of a question with no associated file
From the Forums Explore page, clicking on a question independent of a
file gave you a critical error, since FileSideView component was being
rendered, and it expected many parameters such as the thumbnail of the
file, whereas in reality the file property of the question was null.

To resolve this, FileSideView component is rendered only if the
particular question has a file associated with it.